+++ Sources/Sentry/SentryTracer.m
@@ -145,7 +145,6 @@ static BOOL appStartMeasurementRead;
     }
 
 #if SENTRY_HAS_UIKIT
-    appStartMeasurement = [self getAppStartMeasurement];
     viewNames = [SentryDependencyContainer.sharedInstance.application relevantViewControllersNames];
 #endif // SENTRY_HAS_UIKIT
 
@@ -497,6 +496,8 @@ static BOOL appStartMeasurementRead;
         [super finishWithStatus:_finishStatus];
     }
 #if SENTRY_HAS_UIKIT
+    appStartMeasurement = [self getAppStartMeasurement];
+
     if (appStartMeasurement != nil) {
         [self updateStartTime:appStartMeasurement.appStartTimestamp];
     }