+++ Sources/Sentry/include/SentryAppStartTracker.h
@@ -6,6 +6,7 @@
 
 @class SentryDispatchQueueWrapper;
 @class SentryAppStateManager;
+@class SentryFramesTracker;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -22,7 +23,9 @@ SENTRY_NO_INIT
 
 - (instancetype)initWithDispatchQueueWrapper:(SentryDispatchQueueWrapper *)dispatchQueueWrapper
                              appStateManager:(SentryAppStateManager *)appStateManager
-              enablePreWarmedAppStartTracing:(BOOL)enablePreWarmedAppStartTracing;
+                               framesTracker:(SentryFramesTracker *)framesTracker
+              enablePreWarmedAppStartTracing:(BOOL)enablePreWarmedAppStartTracing
+                         enablePerformanceV2:(BOOL)enablePerformanceV2;
 
 - (void)start;
 - (void)stop;